extract game bson handlers
