+++ modules/game/src/main/GameDiff.scala
@@ -46,8 +46,8 @@ private[game] object GameDiff {
     d(castleLastMoveTime, _.castleLastMoveTime, CastleLastMoveTime.castleLastMoveTimeBSONHandler.write)
     d(moveTimes, _.moveTimes, (x: Vector[Int]) => ByteArray.ByteArrayBSONHandler.write(BinaryFormat.moveTime write x))
     dOpt(positionHashes, _.positionHashes, w.bytesO)
-    dOpt(clock, _.clock, (o: Option[Clock]) => o map { c => Game clockBSONHandler a.createdAt write (_ => c) })
-    dOpt(checkCount, _.checkCount, (o: CheckCount) => o.nonEmpty option { Game.checkCountWriter write o })
+    dOpt(clock, _.clock, (o: Option[Clock]) => o map { c => BSONHandlers clockBSONHandler a.createdAt write (_ => c) })
+    dOpt(checkCount, _.checkCount, (o: CheckCount) => o.nonEmpty option { BSONHandlers.checkCountWriter write o })
     for (i ← 0 to 1) {
       import Player.BSONFields._
       val name = s"p$i."
